+++ plugin-gradle/src/test/java/com/diffplug/gradle/spotless/ToggleOffOnTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020-2021 DiffPlug
+ * Copyright 2020-2025 DiffPlug
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -17,17 +17,48 @@ package com.diffplug.gradle.spotless;
 
 import java.io.IOException;
 
+import org.gradle.testkit.runner.GradleRunner;
 import org.junit.jupiter.api.Test;
 
-class ToggleOffOnTest extends GradleIntegrationHarness {
+abstract class ToggleOffOnTest extends GradleIntegrationHarness {
+	private boolean useConfigCache;
+
+	ToggleOffOnTest(boolean useConfigCache) {
+		this.useConfigCache = useConfigCache;
+	}
+
+	static class WithConfigCache extends ToggleOffOnTest {
+		WithConfigCache() {
+			super(true);
+		}
+	}
+
+	static class WithoutConfigCache extends ToggleOffOnTest {
+		WithoutConfigCache() {
+			super(false);
+		}
+	}
+
+	@Override
+	public GradleRunner gradleRunner() throws IOException {
+		if (useConfigCache) {
+			setFile("gradle.properties").toLines("org.gradle.unsafe.configuration-cache=true",
+					"org.gradle.configuration-cache=true");
+			return super.gradleRunner().withGradleVersion(GradleVersionSupport.CONFIGURATION_CACHE.version);
+		} else {
+			return super.gradleRunner();
+		}
+	}
+
 	@Test
-	void toggleOffOn() throws IOException {
+	void lowercase() throws IOException {
 		setFile("build.gradle").toLines(
 				"plugins { id 'com.diffplug.spotless' }",
 				"spotless {",
+				"  lineEndings 'UNIX'",
 				"  format 'toLower', {",
 				"    target '**/*.md'",
-				"    custom 'lowercase', { str -> str.toLowerCase() }",
+				"    addStep(com.diffplug.spotless.TestingOnly.lowercase())",
 				"    toggleOffOn()",
 				"  }",
 				"}");
@@ -37,12 +68,50 @@ class ToggleOffOnTest extends GradleIntegrationHarness {
 				"D E F",
 				"spotless:on",
 				"G H I");
-		gradleRunner().withArguments("spotlessApply").build();
+		gradleRunner().withArguments("spotlessApply", "--stacktrace").build();
 		assertFile("test.md").hasLines(
 				"a b c",
 				"spotless:off",
 				"D E F",
 				"spotless:on",
 				"g h i");
+		gradleRunner().withArguments("spotlessApply", "--stacktrace").build();
+		gradleRunner().withArguments("spotlessApply", "--stacktrace").build();
+		gradleRunner().withArguments("spotlessApply", "--stacktrace").build();
+		setFile("test.md").toLines(
+				"A B C",
+				"spotless:off",
+				"D E F",
+				"spotless:on",
+				"G H I");
+		gradleRunner().withArguments("spotlessApply", "--stacktrace").build();
+		assertFile("test.md").hasLines(
+				"a b c",
+				"spotless:off",
+				"D E F",
+				"spotless:on",
+				"g h i");
+	}
+
+	@Test
+	void gjf() throws IOException {
+		setFile("build.gradle").toLines(
+				"plugins {",
+				"    id 'com.diffplug.spotless'",
+				"}",
+				"repositories { mavenCentral() }",
+				"",
+				"spotless {",
+				"    java {",
+				"        target file('test.java')",
+				"        googleJavaFormat('1.17.0')",
+				"        toggleOffOn()",
+				"    }",
+				"}");
+
+		setFile("test.java").toResource("java/googlejavaformat/JavaCodeUnformatted.test");
+		gradleRunner().withArguments("spotlessApply").build();
+		assertFile("test.java").sameAsResource("java/googlejavaformat/JavaCodeFormatted.test");
+		gradleRunner().withArguments("spotlessCheck").build();
 	}
 }